utilizando unit

+++ src/main/kotlin/com/acme/tour/controller/PromocaoController.kt
@@ -3,6 +3,8 @@ package com.acme.tour.controller
 import com.acme.tour.model.Promocao
 import com.acme.tour.service.PromocaoService
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 import java.util.concurrent.ConcurrentHashMap
 
@@ -14,24 +16,45 @@ class PromocaoController {
     lateinit var promocaoService: PromocaoService
 
     @GetMapping
-    fun getAllPromocoes(@RequestParam(required = false, defaultValue = "") localFilter: String) =
-        this.promocaoService.searchByLocal(localFilter)
+    fun getAllPromocoes(@RequestParam(required = false, defaultValue = "") localFilter: String): ResponseEntity<List<Promocao>> {
+        return ResponseEntity(this.promocaoService.searchByLocal(localFilter), HttpStatus.OK)
+    }
 
     @GetMapping("/{id}")
-    fun getPromocaoById(@PathVariable id: Long) = this.promocaoService.getById(id)
+    fun getPromocaoById(@PathVariable id: Long): ResponseEntity<Promocao?> {
+        var promocao = this.promocaoService.getById(id)
+        var status = if(promocao == null) HttpStatus.NOT_FOUND else HttpStatus.OK
+
+        return ResponseEntity(promocao, status)
+    }
 
     @PostMapping
-    fun createPromocao(@RequestBody promocao: Promocao) {
+    fun createPromocao(@RequestBody promocao: Promocao): ResponseEntity<Unit> {
         this.promocaoService.create(promocao)
+        return ResponseEntity(Unit, HttpStatus.CREATED)
     }
 
     @DeleteMapping("/{id}")
-    fun deletePromocao(@PathVariable id: Long) {
-        this.promocaoService.delete(id)
+    fun deletePromocao(@PathVariable id: Long): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if(this.promocaoService.getById(id) != null) {
+            status = HttpStatus.ACCEPTED
+            this.promocaoService.delete(id)
+        }
+
+        return ResponseEntity(Unit, status)
     }
 
     @PutMapping("/{id}")
-    fun updatePromocao(@PathVariable id: Long, @RequestBody promocao: Promocao) {
-        this.promocaoService.update(id, promocao)
+    fun updatePromocao(@PathVariable id: Long, @RequestBody promocao: Promocao): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+
+        if(this.promocaoService.getById(id) != null) {
+            status = HttpStatus.ACCEPTED
+            this.promocaoService.update(id, promocao)
+        }
+
+        return ResponseEntity(Unit, status)
     }
 }
